Theme button added -Seperate js file for layouts created -Fake quotes behave properly in QOTD aswell -Class and ID names improved
