+++ QOTD.css
@@ -12,7 +12,7 @@ a:hover {
 }
 
 .material-symbols-outlined {
-	font-variation-settings: "FILL" 0, "wght" 400, "GRAD" 0, "opsz" 48;
+	font-variation-settings: "FILL" 0, "wght" 700, "GRAD" 0, "opsz" 48;
 }
 #backIcon {
 	margin-right: 10%;
@@ -56,9 +56,7 @@ a:hover {
 }
 .quotecard {
 	text-align: center;
-	display: flex;
 	justify-content: center;
-	align-items: center;
 	flex-direction: column;
 	padding: 1rem 2rem;
 	margin: 0 1.5rem;
