+++ quotes.html
@@ -16,6 +16,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.9.1/gsap.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.9.1/ScrollTrigger.min.js"></script>
     <script src="Scripts/quoteSearch.js" defer></script>
+    <script src="Scripts/quoteLayout.js" defer></script>
     <script src="Scripts/quoteoftheday.js" type="module" defer></script>
 </head>
 
@@ -36,8 +37,8 @@
                 <div class="card titlecard">
                     <h1>Quote Of The Day</h1>
                 </div>
-                <div class="card quotecard">
-                    <div id="cardtext"></div>
+                <div class="card quotecard flex-parent">
+                    <div id="cardtext" class="actuallyQuote"></div>
                     <div id="cardauthor"></div>
                 </div>
             </div>
@@ -46,18 +47,22 @@
 
     <!-- All quotes and searching -->
     <section id="search-section">
+        <!-- Search bar and buttons -->
         <div id="searchbar-container">
             <div id="searchandlabel-wrapper">
                 <label for="search" class="see-through">Search</label>
                 <input type="search" id="search" class="see-through" data-search>
             </div>
             <div id="buttonsContainer">
-                <div class="material-symbols-outlined see-through" id="rowsIcon">
+                <div class="material-symbols-outlined see-through accessibility-buttons" id="rows-icon" >
                     grid_on
                 </div>
+                <div class="material-symbols-outlined see-through accessibility-buttons" id="light-dark-mode" title="Light Mode" >
+                    light_mode
+                </div>
             </div>
-
         </div>
+        <!-- Quotes start -->
         <div class="all-quotes gridLayout" data-cards-container>
             <div id="emptyContainer" class="imageHide">
                 <img src="Pics/lostbot.png" id="emptyImage" title="No Result" alt="No result">
